Simple decomposition
