--- lab_5/Local-chat/Tests/clients_tests.py
@@ -1,14 +1,0 @@
-
-import pytest
-
-from main import RegexProcessor
-
-def test_connect_server():
-    ...
-
-def test_send_msg():
-    ...
-
-
-def test_receive_msg():
-    ...
