--- lab_5/Local-chat/Tests/server_tests.py
@@ -1,11 +1,0 @@
-import pytest
-
-from main import RegexProcessor
-
-def test_handle_client():
-    ...
-
-def test_broadcast():
-    ...
-
-
