--- lab_5/Local-chat/requirements.txt
@@ -1,2 +1,0 @@
-art
-tk
