+++ lab_5/Local_chat/App/Tests/server_test.py
@@ -1,0 +1,11 @@
+# import pytest
+
+# from main import RegexProcessor
+
+# def test_handle_client():
+#     ...
+
+# def test_broadcast():
+#     ...
+
+
