+++ lab_5/Local_chat/main.py
@@ -1,0 +1,6 @@
+
+from App.app import Start
+
+
+if __name__ == "__main__":  
+    Start.main_start()
